fix: Do not crash at startup when client not ready when fetching geolocation webhook
